requirements.txt
Flask
gunicorn
psycopg2-binary
python-dotenv
werkzeug